Extract Swapbox to it's own component
