+++ src/components/Swap/SwapButtons/SwapButtons.tsx
@@ -7,7 +7,6 @@ import { PRICE_IMPACT_HIGH, PRICE_IMPACT_MEDIUM, ROUTABLE_PLATFORM_STYLE } from 
 import { useActiveWeb3React } from '../../../hooks'
 import { ApprovalState } from '../../../hooks/useApproveCallback'
 import { WrapState, WrapType } from '../../../hooks/useWrapCallback'
-import { SwapData } from '../../../pages/Swap'
 import { Field } from '../../../state/swap/types'
 import { useIsExpertMode } from '../../../state/user/hooks'
 import { warningSeverity } from '../../../utils/prices'
@@ -18,6 +17,7 @@ import Column from '../../Column'
 import { Loader } from '../../Loader'
 import ProgressSteps from '../../ProgressSteps'
 import { AutoRow, RowBetween } from '../../Row'
+import { SwapData } from '../../Swapbox/Swapbox'
 import { ErrorBox } from './SwabButtonsError'
 import { SwapButton, SwapLoadingButton } from './SwapButton'
