+++ src/pages/Swap/index.tsx
@@ -1,50 +1,24 @@
-import { CoWTrade, Currency, CurrencyAmount, JSBI, RoutablePlatform, Token, Trade } from '@swapr/sdk'
+import { CoWTrade, RoutablePlatform, Token } from '@swapr/sdk'
 
-// Landing Page Imports
 import './../../theme/landingPageTheme/stylesheet.css'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
-import { Box, Flex } from 'rebass'
+import { useCallback, useMemo, useState } from 'react'
+import { Flex } from 'rebass'
 import styled from 'styled-components'
 
-import { ReactComponent as SwapIcon } from '../../assets/images/swap-icon.svg'
 import { ButtonGroup, ButtonGroupOption } from '../../components/ButtonGroup/ButtonGroup'
-import { AutoColumn } from '../../components/Column'
-import { CurrencyInputPanel } from '../../components/CurrencyInputPanel'
-import { SwapPoolTabs } from '../../components/NavigationTabs'
 import { SimpleChartContainer } from '../../components/SimpleChart/SimpleChartContainer'
 import AdvancedSwapDetailsDropdown from '../../components/Swap/AdvancedSwapDetailsDropdown'
-import confirmPriceImpactWithoutFee from '../../components/Swap/confirmPriceImpactWithoutFee'
-import ConfirmSwapModal from '../../components/Swap/ConfirmSwapModal'
-import { ArrowWrapper, SwitchTokensAmountsContainer, Wrapper } from '../../components/Swap/styleds'
-import SwapButtons from '../../components/Swap/SwapButtons'
 import { Tabs } from '../../components/Swap/Tabs'
-import { TradeDetails } from '../../components/Swap/TradeDetails'
+import { Swapbox } from '../../components/Swapbox/Swapbox'
 import TokenWarningModal from '../../components/TokenWarningModal'
 import { useActiveWeb3React } from '../../hooks'
 import { useAllTokens, useCurrency } from '../../hooks/Tokens'
-import { ApprovalState, useApproveCallbackFromTrade } from '../../hooks/useApproveCallback'
-import { useSwapCallback } from '../../hooks/useSwapCallback'
 import { useTargetedChainIdFromUrl } from '../../hooks/useTargetedChainIdFromUrl'
-import { useHigherUSDValue } from '../../hooks/useUSDValue'
-import { useWrapCallback, WrapState, WrapType } from '../../hooks/useWrapCallback'
-import {
-  useDefaultsFromURLSearch,
-  useDerivedSwapInfo,
-  useSwapActionHandlers,
-  useSwapState,
-} from '../../state/swap/hooks'
+import { useWrapCallback, WrapType } from '../../hooks/useWrapCallback'
+import { useDefaultsFromURLSearch, useDerivedSwapInfo, useSwapState } from '../../state/swap/hooks'
 import { Field } from '../../state/swap/types'
-import {
-  useAdvancedSwapDetails,
-  useIsExpertMode,
-  useUpdateSelectedChartOption,
-  useUserSlippageTolerance,
-} from '../../state/user/hooks'
+import { useAdvancedSwapDetails, useUpdateSelectedChartOption } from '../../state/user/hooks'
 import { ChartOptions } from '../../state/user/reducer'
-import { computeFiatValuePriceImpact } from '../../utils/computeFiatValuePriceImpact'
-import { maxAmountSpend } from '../../utils/maxAmountSpend'
-import { computeTradePriceBreakdown, warningSeverity } from '../../utils/prices'
-import AppBody from '../AppBody'
 import BlogNavigation from './../../components/LandingPageComponents/BlogNavigation'
 import CommunityBanner from './../../components/LandingPageComponents/CommunityBanner'
 import CommunityLinks from './../../components/LandingPageComponents/CommunityLinks'
@@ -53,20 +27,6 @@ import Footer from './../../components/LandingPageComponents/layout/Footer'
 import Hero from './../../components/LandingPageComponents/layout/Hero'
 import Stats from './../../components/LandingPageComponents/Stats'
 import Timeline from './../../components/LandingPageComponents/Timeline'
-
-export type SwapData = {
-  showConfirm: boolean
-  tradeToConfirm: Trade | undefined
-  attemptingTxn: boolean
-  swapErrorMessage: string | undefined
-  txHash: string | undefined
-}
-
-const SwitchIconContainer = styled.div`
-  height: 0;
-  position: relative;
-  width: 100%;
-`
 
 const AppBodyContainer = styled.section`
   display: flex;
@@ -92,18 +52,11 @@ const LandingBodyContainer = styled.section`
   width: calc(100% + 32px) !important;
 `
 
-export enum CoWTradeState {
-  UNKNOWN, // default
-  WRAP,
-  APPROVAL,
-  SWAP,
-}
-
 export default function Swap() {
   const loadedUrlParams = useDefaultsFromURLSearch()
   const [platformOverride, setPlatformOverride] = useState<RoutablePlatform | null>(null)
   const allTokens = useAllTokens()
-  const [showAdvancedSwapDetails, setShowAdvancedSwapDetails] = useAdvancedSwapDetails()
+  const [showAdvancedSwapDetails] = useAdvancedSwapDetails()
 
   // token warning stuff
   const [loadedInputCurrency, loadedOutputCurrency] = [
@@ -128,259 +81,26 @@ export default function Swap() {
   const { chainId } = useActiveWeb3React()
   const [selectedChartOption, setselectedChartOption] = useUpdateSelectedChartOption()
 
-  // for expert mode
-  const isExpertMode = useIsExpertMode()
-
-  // get custom setting values for user
-  const allowedSlippage = useUserSlippageTolerance()
-
   // swap state
-  const { independentField, typedValue, recipient } = useSwapState()
+  const { typedValue } = useSwapState()
 
   const {
     trade: potentialTrade,
     allPlatformTrades,
-    currencyBalances,
-    parsedAmount,
     currencies,
-    inputError: swapInputError,
     loading,
   } = useDerivedSwapInfo(platformOverride || undefined)
 
-  // For GPv2 trades, have a state which holds: approval status (handled by useApproveCallback), and
-  // wrap status(use useWrapCallback and a state variable)
-  const [gnosisProtocolTradeState, setGnosisProtocolState] = useState(CoWTradeState.UNKNOWN)
-  const {
-    wrapType,
-    execute: onWrap,
-    inputError: wrapInputError,
-    wrapState,
-    setWrapState,
-  } = useWrapCallback(
+  const { wrapType } = useWrapCallback(
     currencies.INPUT,
     currencies.OUTPUT,
     potentialTrade instanceof CoWTrade,
     potentialTrade?.inputAmount?.toSignificant(6) ?? typedValue
   )
 
-  const bestPricedTrade = allPlatformTrades?.[0]
   const showWrap = wrapType !== WrapType.NOT_APPLICABLE && !(potentialTrade instanceof CoWTrade)
 
   const trade = showWrap ? undefined : potentialTrade
-
-  //GPv2 is falling in the true case, not very useful for what I think I need
-  const parsedAmounts = showWrap
-    ? {
-        [Field.INPUT]: parsedAmount,
-        [Field.OUTPUT]: parsedAmount,
-      }
-    : {
-        [Field.INPUT]: independentField === Field.INPUT ? parsedAmount : trade?.inputAmount,
-        [Field.OUTPUT]: independentField === Field.OUTPUT ? parsedAmount : trade?.outputAmount,
-      }
-
-  const { onSwitchTokens, onCurrencySelection, onUserInput } = useSwapActionHandlers()
-
-  const dependentField: Field = independentField === Field.INPUT ? Field.OUTPUT : Field.INPUT
-
-  const handleTypeInput = useCallback(
-    (value: string) => {
-      onUserInput(Field.INPUT, value)
-    },
-    [onUserInput]
-  )
-  const handleTypeOutput = useCallback(
-    (value: string) => {
-      onUserInput(Field.OUTPUT, value)
-    },
-    [onUserInput]
-  )
-
-  // modal and loading
-  const [{ showConfirm, tradeToConfirm, swapErrorMessage, attemptingTxn, txHash }, setSwapState] = useState<SwapData>({
-    showConfirm: false,
-    tradeToConfirm: undefined,
-    attemptingTxn: false,
-    swapErrorMessage: undefined,
-    txHash: undefined,
-  })
-
-  const formattedAmounts = {
-    [independentField]: typedValue,
-    [dependentField]: showWrap
-      ? parsedAmounts[independentField]?.toExact() ?? ''
-      : parsedAmounts[dependentField]?.toSignificant(6) ?? '',
-  }
-
-  const userHasSpecifiedInputOutput = Boolean(
-    currencies[Field.INPUT] && currencies[Field.OUTPUT] && parsedAmounts[independentField]?.greaterThan(JSBI.BigInt(0))
-  )
-
-  // check whether the user has approved the router on the input token
-  const [approval, approveCallback] = useApproveCallbackFromTrade(trade /* allowedSlippage */)
-
-  // check if user has gone through approval process, used to show two step buttons, reset on token change
-  const [approvalsSubmitted, setApprovalsSubmitted] = useState<boolean[]>([])
-
-  const currentTradeIndex = useMemo(() => {
-    if (allPlatformTrades) {
-      return (
-        allPlatformTrades.findIndex(
-          trade => trade && platformOverride && trade.platform.name === platformOverride.name
-        ) || 0
-      )
-    } else {
-      return 0
-    }
-  }, [platformOverride, allPlatformTrades])
-
-  // mark when a user has submitted an approval, reset onTokenSelection for input field
-  useEffect(() => {
-    if (approval === ApprovalState.PENDING) {
-      const newArray = [...approvalsSubmitted]
-      newArray[currentTradeIndex] = true
-      setApprovalsSubmitted(newArray)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [approval])
-
-  // Listen for changes on wrapState
-  useEffect(() => {
-    // watch GPv2
-    if (wrapState === WrapState.WRAPPED) {
-      if (approval === ApprovalState.APPROVED) setGnosisProtocolState(CoWTradeState.SWAP)
-      else setGnosisProtocolState(CoWTradeState.APPROVAL)
-    }
-  }, [wrapState, approval, trade])
-
-  const maxAmountInput: CurrencyAmount | undefined = maxAmountSpend(currencyBalances[Field.INPUT], chainId)
-  const maxAmountOutput: CurrencyAmount | undefined = maxAmountSpend(currencyBalances[Field.OUTPUT], chainId, false)
-
-  // the callback to execute the swap
-  const { callback: swapCallback, error: swapCallbackError } = useSwapCallback({
-    trade,
-    allowedSlippage,
-    recipientAddressOrName: recipient,
-  })
-
-  const { priceImpactWithoutFee } = computeTradePriceBreakdown(trade)
-
-  const handleSwap = useCallback(() => {
-    if (priceImpactWithoutFee && !confirmPriceImpactWithoutFee(priceImpactWithoutFee)) {
-      return
-    }
-    if (!swapCallback) {
-      return
-    }
-    setSwapState({
-      attemptingTxn: true,
-      tradeToConfirm,
-      showConfirm,
-      swapErrorMessage: undefined,
-      txHash: undefined,
-    })
-    swapCallback()
-      .then(hash => {
-        setSwapState({
-          attemptingTxn: false,
-          tradeToConfirm,
-          showConfirm,
-          swapErrorMessage: undefined,
-          txHash: hash,
-        })
-        //reset states, in case we want to operate again
-        setGnosisProtocolState(CoWTradeState.UNKNOWN)
-        setWrapState && setWrapState(WrapState.UNKNOWN)
-      })
-      .catch(error => {
-        setSwapState({
-          attemptingTxn: false,
-          tradeToConfirm,
-          showConfirm,
-          swapErrorMessage: error.message,
-          txHash: undefined,
-        })
-        setGnosisProtocolState(CoWTradeState.UNKNOWN)
-      })
-  }, [priceImpactWithoutFee, swapCallback, tradeToConfirm, showConfirm, setWrapState])
-
-  // warnings on slippage
-  const priceImpactSeverity = warningSeverity(priceImpactWithoutFee)
-
-  // show approve flow when: no error on inputs, not approved or pending, or approved in current session
-  // never show if price impact is above threshold in non expert mode
-  const showApproveFlow =
-    !swapInputError &&
-    (approval === ApprovalState.NOT_APPROVED ||
-      approval === ApprovalState.PENDING ||
-      (approvalsSubmitted[currentTradeIndex] && approval === ApprovalState.APPROVED)) &&
-    !(priceImpactSeverity > 3 && !isExpertMode)
-
-  const handleConfirmDismiss = useCallback(() => {
-    setSwapState({
-      showConfirm: false,
-      tradeToConfirm,
-      attemptingTxn,
-      swapErrorMessage,
-      txHash,
-    })
-    // if there was a tx hash, we want to clear the input
-    if (txHash) {
-      onUserInput(Field.INPUT, '')
-      onUserInput(Field.OUTPUT, '')
-    }
-  }, [attemptingTxn, onUserInput, swapErrorMessage, tradeToConfirm, txHash])
-
-  const handleAcceptChanges = useCallback(() => {
-    setSwapState({
-      tradeToConfirm: trade,
-      swapErrorMessage,
-      txHash,
-      attemptingTxn,
-      showConfirm,
-    })
-  }, [attemptingTxn, showConfirm, swapErrorMessage, trade, txHash])
-
-  const handleInputSelect = useCallback(
-    (inputCurrency: Currency) => {
-      setPlatformOverride(null) // reset platform override, since best prices might be on a different platform
-      setApprovalsSubmitted([]) // reset 2 step UI for approvals
-      onCurrencySelection(Field.INPUT, inputCurrency)
-    },
-    [onCurrencySelection]
-  )
-
-  const handleMaxInput = useCallback(
-    (fieldInput: string) => () => {
-      maxAmountInput && fieldInput === Field.INPUT && onUserInput(Field.INPUT, maxAmountInput.toExact())
-      maxAmountOutput && fieldInput === Field.OUTPUT && onUserInput(Field.OUTPUT, maxAmountOutput.toExact())
-    },
-    [maxAmountInput, maxAmountOutput, onUserInput]
-  )
-
-  const handleOutputSelect = useCallback(
-    (outputCurrency: Currency) => {
-      setPlatformOverride(null) // reset platform override, since best prices might be on a different platform
-      onCurrencySelection(Field.OUTPUT, outputCurrency)
-    },
-    [onCurrencySelection]
-  )
-
-  const { fiatValueInput, fiatValueOutput, isFallbackFiatValueInput, isFallbackFiatValueOutput } = useHigherUSDValue({
-    inputCurrencyAmount: parsedAmounts[Field.INPUT],
-    outputCurrencyAmount: parsedAmounts[Field.OUTPUT],
-  })
-
-  const priceImpact = useMemo(
-    () => computeFiatValuePriceImpact(fiatValueInput, fiatValueOutput),
-    [fiatValueInput, fiatValueOutput]
-  )
-
-  const isInputPanelDisabled =
-    (gnosisProtocolTradeState === CoWTradeState.APPROVAL ||
-      gnosisProtocolTradeState === CoWTradeState.SWAP ||
-      wrapState === WrapState.PENDING) &&
-    trade instanceof CoWTrade
 
   const hasBothCurrenciesInput = currencies[Field.INPUT] && currencies[Field.OUTPUT]
 
@@ -410,125 +130,25 @@ export default function Swap() {
               flexDirection={['column-reverse', 'row']}
             >
               <Tabs />
-              <Box mb={[3, 0]}>
-                <ButtonGroup>
-                  <ButtonGroupOption
-                    disabled={!hasBothCurrenciesInput}
-                    active={selectedChartOption === ChartOptions.SIMPLE_CHART}
-                    onClick={() => setselectedChartOption(ChartOptions.SIMPLE_CHART)}
-                  >
-                    Chart
-                  </ButtonGroupOption>
-                  <ButtonGroupOption
-                    disabled={!hasBothCurrenciesInput}
-                    active={selectedChartOption === ChartOptions.OFF}
-                    onClick={() => setselectedChartOption(ChartOptions.OFF)}
-                  >
-                    Off
-                  </ButtonGroupOption>
-                </ButtonGroup>
-              </Box>
+
+              <ButtonGroup mb={[3, 0]}>
+                <ButtonGroupOption
+                  disabled={!hasBothCurrenciesInput}
+                  active={selectedChartOption === ChartOptions.SIMPLE_CHART}
+                  onClick={() => setselectedChartOption(ChartOptions.SIMPLE_CHART)}
+                >
+                  Chart
+                </ButtonGroupOption>
+                <ButtonGroupOption
+                  disabled={!hasBothCurrenciesInput}
+                  active={selectedChartOption === ChartOptions.OFF}
+                  onClick={() => setselectedChartOption(ChartOptions.OFF)}
+                >
+                  Off
+                </ButtonGroupOption>
+              </ButtonGroup>
             </Flex>
-            <AppBody tradeDetailsOpen={!!trade}>
-              <SwapPoolTabs active="swap" />
-              <Wrapper id="swap-page">
-                <ConfirmSwapModal
-                  isOpen={showConfirm}
-                  trade={trade}
-                  originalTrade={tradeToConfirm}
-                  onAcceptChanges={handleAcceptChanges}
-                  attemptingTxn={attemptingTxn}
-                  txHash={txHash}
-                  recipient={recipient}
-                  allowedSlippage={allowedSlippage}
-                  onConfirm={handleSwap}
-                  swapErrorMessage={swapErrorMessage}
-                  onDismiss={handleConfirmDismiss}
-                />
-
-                <AutoColumn gap="12px">
-                  <AutoColumn gap="3px">
-                    <CurrencyInputPanel
-                      value={formattedAmounts[Field.INPUT]}
-                      currency={currencies[Field.INPUT]}
-                      onUserInput={handleTypeInput}
-                      onMax={handleMaxInput(Field.INPUT)}
-                      onCurrencySelect={handleInputSelect}
-                      otherCurrency={currencies[Field.OUTPUT]}
-                      fiatValue={fiatValueInput}
-                      isFallbackFiatValue={isFallbackFiatValueInput}
-                      maxAmount={maxAmountInput}
-                      showCommonBases
-                      disabled={isInputPanelDisabled}
-                      id="swap-currency-input"
-                    />
-                    <SwitchIconContainer>
-                      <SwitchTokensAmountsContainer
-                        onClick={() => {
-                          setApprovalsSubmitted([]) // reset 2 step UI for approvals
-                          onSwitchTokens()
-                        }}
-                      >
-                        <ArrowWrapper
-                          clickable={!loading}
-                          data-testid="switch-tokens-button"
-                          className={loading ? 'rotate' : ''}
-                        >
-                          <SwapIcon />
-                        </ArrowWrapper>
-                      </SwitchTokensAmountsContainer>
-                    </SwitchIconContainer>
-                    <CurrencyInputPanel
-                      value={formattedAmounts[Field.OUTPUT]}
-                      onUserInput={handleTypeOutput}
-                      onMax={handleMaxInput(Field.OUTPUT)}
-                      currency={currencies[Field.OUTPUT]}
-                      onCurrencySelect={handleOutputSelect}
-                      otherCurrency={currencies[Field.INPUT]}
-                      fiatValue={fiatValueOutput}
-                      priceImpact={priceImpact}
-                      isFallbackFiatValue={isFallbackFiatValueOutput}
-                      maxAmount={maxAmountOutput}
-                      showCommonBases
-                      disabled={isInputPanelDisabled}
-                      id="swap-currency-output"
-                    />
-                  </AutoColumn>
-
-                  <TradeDetails
-                    show={!showWrap}
-                    loading={loading}
-                    trade={trade}
-                    bestPricedTrade={bestPricedTrade}
-                    showAdvancedSwapDetails={showAdvancedSwapDetails}
-                    setShowAdvancedSwapDetails={setShowAdvancedSwapDetails}
-                    recipient={recipient}
-                  />
-                  <SwapButtons
-                    wrapInputError={wrapInputError}
-                    showApproveFlow={showApproveFlow}
-                    userHasSpecifiedInputOutput={userHasSpecifiedInputOutput}
-                    approval={approval}
-                    setSwapState={setSwapState}
-                    priceImpactSeverity={priceImpactSeverity}
-                    swapCallbackError={swapCallbackError}
-                    wrapType={wrapType}
-                    approvalSubmitted={approvalsSubmitted[currentTradeIndex]}
-                    currencies={currencies}
-                    trade={trade}
-                    swapInputError={swapInputError}
-                    swapErrorMessage={swapErrorMessage}
-                    loading={loading}
-                    onWrap={onWrap}
-                    approveCallback={approveCallback}
-                    handleSwap={handleSwap}
-                    handleInputSelect={handleInputSelect}
-                    wrapState={wrapState}
-                    setWrapState={setWrapState}
-                  />
-                </AutoColumn>
-              </Wrapper>
-            </AppBody>
+            <Swapbox />
             {showAdvancedSwapDetails && (
               <AdvancedSwapDetailsDropdown
                 isLoading={loading}
